move `gestureProps` to types.ts
