+++ src/motion/features/gestures.ts
@@ -1,22 +1,8 @@
 import { MotionProps } from "../types"
-import { GestureHandlers, useGestures } from "../../gestures"
+import { useGestures } from "../../gestures"
+import { gestureProps } from "../../gestures/types"
 import { FeatureProps, MotionFeature } from "./types"
 import { makeRenderlessComponent } from "../utils/make-renderless-component"
-
-export const gestureProps: Array<keyof GestureHandlers> = [
-    "onPan",
-    "onPanStart",
-    "onPanEnd",
-    "onPanSessionStart",
-    "onTap",
-    "onTapStart",
-    "onTapCancel",
-    "onHoverStart",
-    "onHoverEnd",
-    "whileFocus",
-    "whileTap",
-    "whileHover",
-]
 
 const GestureComponent = makeRenderlessComponent(
     ({ visualElement, ...props }: FeatureProps) => {
